Modify:modify the PEM API
